add time based filter

+++ remove-databases/main.py
@@ -2,6 +2,12 @@ import os
 import yaml
 import pymysql
 
+web_root = "/var/lib/docker/workspace/infinity/html/OHRMStandalone"
+db_host = "172.40.0.102"
+db_user = "root"
+root_password = "p2835GmVeBpbPLT"
+soft_deletion = False
+
 
 def get_legitimate_db_list(webroot):
     db_array = []
@@ -32,7 +38,7 @@ def get_legitimate_db_list(webroot):
 
 
 def get_total_db_list(db_host, root_password):
-    db_connection = pymysql.connect(db_host, "root", root_password)
+    db_connection = pymysql.connect(db_host, db_user, root_password)
     db_pointer = db_connection.cursor()
     db_pointer.execute("show databases")
     result = db_pointer.fetchall()
@@ -40,37 +46,58 @@ def get_total_db_list(db_host, root_password):
     return result
 
 
-def database_filter(legitimate_list, database):
-    if database not in legitimate_list:
-        return True
-    return False
+def database_filter_by_date(time_period):
+    sql_query = "SELECT table_schema,MIN(create_time) create_time FROM information_schema.tables Group by " \
+                "TABLE_SCHEMA having create_time < DATE(NOW()) - INTERVAL " + time_period + "DAY Order by create_time " \
+                                                                                            "desc;"
+    db_connection = pymysql.connect(db_host, db_user, root_password)
+    db_pointer = db_connection.cursor()
+    db_pointer.execute(sql_query)
+    result = db_pointer.fetchall()
+    db_connection.close()
+    final_list = []
+    if result.count() <= 0:
+        return []
+
+    result_list = list(result)
+    for x in result_list:
+        final_list.append(x[0])
+
+    return final_list
+
+
+def database_filter(database):
+    legitimate_list = get_legitimate_db_list(web_root)
+    old_databases = database_filter_by_date(7)
+
+    if database in legitimate_list:
+        return False
+    if database not in old_databases:
+        return False
+    return True
 
 
 def delete_databases(db_name, db_host, root_password):
-    exclusion_list = ["information_schema", "mysql", "performance_schema", "requestdesk_centraldb_11_release",
-                      "uat_rd_dummy_db"]
+    exclusion_list = ["information_schema", "mysql", "performance_schema"]
     if db_name not in exclusion_list:
-        db_connection = pymysql.connect(db_host, "root", root_password)
+        db_connection = pymysql.connect(db_host, db_user, root_password)
         db_pointer = db_connection.cursor()
-        db_pointer.execute("drop database `" + db_name + "`;")
+        if soft_deletion:
+            print(db_name)
+        else:
+            db_pointer.execute("drop database `" + db_name + "`;")
         result = db_pointer.fetchone()
         db_connection.close()
-        print(db_name)
+
         return result
     return False
 
 
 def main():
-    db_host = "172.40.0.102"
-    root_password = "p2835GmVeBpbPLT"
-    web_root = "/var/lib/docker/workspace/infinity/html/OHRMStandalone"
-
-    database_array = get_legitimate_db_list(web_root)
-    print(database_array)
     total_db_list = get_total_db_list(db_host, root_password)
     for db in total_db_list:
         db_name = db[0]
-        if database_filter(database_array, db_name):
+        if database_filter(db_name):
             delete_databases(db_name, db_host, root_password)
     return True
